requirements.txt
streamlit
langchain
langchain-community
langchain-groq
typing-extensions
python-dotenv
pymysql
pandas
matplotlib